Removed set of known ids and tidied up data sizes

+++ src/ecs/core/id.c
@@ -17,7 +17,7 @@ CE_Result CE_Id_setKind(INOUT CE_Id* id, CE_IdKind kind)
     return CE_OK;
 }
 
-CE_Result CE_Id_setUniqueId(INOUT CE_Id* id, uint32_t uniqueId)
+CE_Result CE_Id_setUniqueId(INOUT CE_Id* id, uint16_t uniqueId)
 {
     if (!id) return CE_ERROR;
     if (uniqueId > CE_ID_MASK_UNIQUE) return CE_ERROR;
@@ -27,7 +27,7 @@ CE_Result CE_Id_setUniqueId(INOUT CE_Id* id, uint32_t uniqueId)
     return CE_OK;
 }
 
-CE_Result CE_Id_setGeneration(INOUT CE_Id* id, uint32_t generation)
+CE_Result CE_Id_setGeneration(INOUT CE_Id* id, uint8_t generation)
 {
     if (!id) return CE_ERROR;
     if (generation > CE_ID_MASK_GENERATION) return CE_ERROR;
@@ -63,7 +63,7 @@ CE_Result CE_Id_setRelationshipTypeId(INOUT CE_Id* id, CE_TypeId typeId)
     return CE_OK;
 }
 
-CE_Result CE_Id_make(IN CE_IdKind kind, IN CE_TypeId typeId, IN uint32_t generation, IN uint32_t uniqueId, OUT CE_Id* out)
+CE_Result CE_Id_make(IN CE_IdKind kind, IN CE_TypeId typeId, IN uint8_t generation, IN uint16_t uniqueId, OUT CE_Id* out)
 {
     if (!out) return CE_ERROR;
